[build] Check for libXext
It seems GNU Gold (the new linker) either behaves as if with
--no-undefined by default, or has some issue preventing it from not
doing that when instructed to (I'm not sure if this actually
happens). In any case, clutter uses the Xshm extensions, but does not
link to libXext, which makes gold unhappy.

Based on a patch by: Xan Lopez <xan@gnome.org>

Signed-off-by: Emmanuele Bassi <ebassi@linux.intel.com>